S3/Minio tests: Set USE_S3 correctly
As @valeriupredoi pointed out, the S3 integration tests (not the S3
exploratory tests) added in #97 are not using S3 storage. This is
because the GitHub Actions workflow writes USE_S3 = True to config.py in
the repo's root directory, rather than activestorage/config.py. The file
is therefore ignored.

This change uses the correct file name.

Fixes: #106