Parallelise chunk processing
Previously, each storage chunk was processed sequentially. This is
functional, but not performant, particularly when using Reductionist
(S3), due to the sequential round trip latencies of requests. This means
we are not taking advantage of the concurrency provided by Reductionist.

This change switches to use a threadpool executor to process chunks.
This is unlikely to be as performant as asyncio, but still provides a
significant improvement without requiring an async runtime.

Some basic throttling of requests is provided by using a threadpool of a
configurable size. This works well when using a single Active object,
but may need refinement if multiple Active objects are in use at the
same time to avoid overloading the Reductionist server.

In order to benefit from connection reuse, we switch to using a requests
Session object that is shared by all threads in the pool.

This change also adds support for specifying a CA certificate to use
when making requests to the Reductionist API.

Closes #121