Use a temporary file for Zarr group JSON
When using multiple netCDF files with the same names, the Zarr group
JSON file would previously not be overwritten after it was first
written. This would lead to subsequent uses potentially using an invalid
Zarr group metadata file.

This change switches to use a temporary file to store the Zarr group
metadata.  This should not be a problem because the Zarr datasource is
cached in the Active object as the _zds member between operations.

Closes #134