Always load netCDF metadata
Since byte order support was added in #132, we capture the dtype from
netCDF metadata and use it in preference to the Zarr dtype when using
the S3 storage backend.

In the case where missing data values are passed into the Active
constructor, we do not load netCDF metadata, and therefore do not
capture the dtype. This leads to the following error with S3 storage:

  AttributeError: 'Active' object has no attribute '_dtype'

This change switches to always load the metadata so that we can capture
the dtype.  This has the downside that it is not necessary when using
local storage and missing data has been specified.

This allows us to enable the test_daily_data_masked test for S3.

Closes #137